Move AdditionalCostsToPlay to common/.
This will help put rendering in the hands of the UI.
