+++ src/server/cards/requirements/CardRequirements.ts
@@ -2,7 +2,8 @@ import {RequirementType} from '../../../common/cards/RequirementType';
 import {Tag} from '../../../common/cards/Tag';
 import {IPlayer} from '../../IPlayer';
 import {CardName} from '../../../common/cards/CardName';
-import {CardRequirement, AdditionalCostsToPlay} from './CardRequirement';
+import {CardRequirement} from './CardRequirement';
+import {AdditionalCostsToPlay} from '../../../common/cards/Types';
 import {ChairmanRequirement} from './ChairmanRequirement';
 import {CitiesRequirement} from './CitiesRequirement';
 import {ColoniesRequirement} from './ColoniesRequirement';
